lr10/cinema_project/requirements.txt
psycopg2-binary==2.9.9
faker==24.8.0
tabulate==0.9.0
python-dotenv==1.0.0
Django>=4.2,<5.0
Faker>=19.0.0